+++ packages/expo-sensors/mocks/ExponentDeviceMotion.ts
@@ -9,4 +9,6 @@ export async function isAvailableAsync(): Promise<any> {}
 
 export async function setUpdateInterval(intervalMs: number): Promise<any> {}
 
-export const Gravity = 9.80665;
+export async function getPermissionsAsync(): Promise<any> {}
+
+export async function requestPermissionsAsync(): Promise<any> {}